Allow ChannelHandler to buffer outbound data and influence Channel writability
Motivation:

In netty 4.1.x we used some hacks to be able to influence Channel writability when outbound data was buffered in a ChannelHandler. While this worked it made things kind of complicated and tied heavily to internals. As buffering outbound data is something that users might need to do as well we should better make it part of the "API".

Modifications:

- Add ChannelHandler.pendingOutboundBytes() which allows handles to signal how much data is buffered and so influence channel writability.
- Rewrite internals to make use of the new method exposed by ChannelHandler.
- Simplify ChannelOutboundBuffer and remove dependency on Channel.

Result:

Less dependency on internals and public API to influence Channel writability from within a ChannelHandler